requirements.txt
flask==3.0.0
flask-cors==4.0.0
trafilatura==1.12.0
openai==1.51.0
python-dotenv==1.0.0
requests==2.31.0
pydantic[email]==2.5.0
supabase==2.8.1
gotrue==2.8.1
PyJWT==2.8.0
colorama==0.4.6
beautifulsoup4==4.12.3
readability-lxml==0.8.1
PyYAML==6.0.1

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Code quality tools (dev only)
ruff==0.6.5
black==24.8.0 